+++ simulator_tests/__init__.py
@@ -7,7 +7,7 @@ Each test is in its own file for better organization and maintainability.
 
 from .base_test import BaseSimulatorTest
 from .test_basic_conversation import BasicConversationTest
-from .test_codereviewworkflow_validation import CodeReviewWorkflowValidationTest
+from .test_codereview_validation import CodeReviewValidationTest
 from .test_consensus_conversation import TestConsensusConversation
 from .test_consensus_stance import TestConsensusStance
 from .test_consensus_three_models import TestConsensusThreeModels
@@ -40,7 +40,7 @@ from .test_xai_models import XAIModelsTest
 # Test registry for dynamic loading
 TEST_REGISTRY = {
     "basic_conversation": BasicConversationTest,
-    "codereviewworkflow_validation": CodeReviewWorkflowValidationTest,
+    "codereview_validation": CodeReviewValidationTest,
     "content_validation": ContentValidationTest,
     "per_tool_deduplication": PerToolDeduplicationTest,
     "cross_tool_continuation": CrossToolContinuationTest,
@@ -73,7 +73,7 @@ TEST_REGISTRY = {
 __all__ = [
     "BaseSimulatorTest",
     "BasicConversationTest",
-    "CodeReviewWorkflowValidationTest",
+    "CodeReviewValidationTest",
     "ContentValidationTest",
     "PerToolDeduplicationTest",
     "CrossToolContinuationTest",
