+++ tools/__init__.py
@@ -5,7 +5,6 @@ Tool implementations for Zen MCP Server
 from .analyze import AnalyzeTool
 from .chat import ChatTool
 from .codereview import CodeReviewTool
-from .codereviewworkflow import CodeReviewWorkflowTool
 from .consensus import ConsensusTool
 from .debug import DebugIssueTool
 from .listmodels import ListModelsTool
@@ -19,7 +18,6 @@ from .tracer import TracerTool
 __all__ = [
     "ThinkDeepTool",
     "CodeReviewTool",
-    "CodeReviewWorkflowTool",
     "DebugIssueTool",
     "AnalyzeTool",
     "ChatTool",
